chore: address pr feedback
- use tailwind arbitrary variants and remove css plugin dep and css file
- add SelectorForm.tsx and export from index.tsx
- use nullish coalescing for var assignment
