+++ packages/javascript/bh-shared-ui/package.json
@@ -83,7 +83,6 @@
         "react-dom": "^18.2.0",
         "rollup": "^4.13.2",
         "rollup-plugin-delete": "^2.0.0",
-        "rollup-plugin-import-css": "^3.5.8",
         "tailwindcss": "^3.4.17",
         "typescript": "^5.1.6",
         "vite": "^5.4.12",
