--- packages/javascript/bh-shared-ui/src/views/TierManagement/Cypher/cypher.css
@@ -1,17 +1,0 @@
-.dark .cm-editor {
-  background-color: transparent;
-}
-
-.dark .cm-editor .cm-gutters {
-  border-color: transparent;
-  background-color: transparent;
-}
-
-.light .cm-editor {
-  background-color: transparent;
-}
-
-.light .cm-editor .cm-gutters {
-  border-color: transparent;
-  background-color: transparent;
-}
